Renaming transactions -> requests
