+++ signer/web/src/components/Requests/index.js
@@ -1,0 +1,1 @@
+export default from './Requests';
