--- signer/web/src/components/Transactions/index.js
@@ -1,1 +1,0 @@
-export default from './Transactions';
