--- signer/web/src/containers/TransactionsPage/index.js
@@ -1,1 +1,0 @@
-export default from './TransactionsPage';
